+++ tests/attributes/test_model.py
@@ -7,7 +7,13 @@ import sheraf
 import tests
 
 
-class Submodel(tests.UUIDAutoModel):
+class Submodel1(sheraf.Model):
+    table = "submodel1"
+    name = sheraf.SimpleAttribute()
+
+
+class Submodel2(sheraf.Model):
+    table = "submodel2"
     name = sheraf.SimpleAttribute()
 
 
@@ -18,24 +24,24 @@ utc_now = datetime.datetime(2016, 12, 31, 23, 59, 59, 5000)
 @pytest.mark.parametrize(
     "model",
     [
-        Submodel,
-        "{}.{}".format(Submodel.__module__, Submodel.__name__),
-        "{}.{}".format(Submodel.__module__, Submodel.__name__).encode("utf-8"),
+        Submodel1,
+        "{}.{}".format(Submodel1.__module__, Submodel1.__name__),
+        "{}.{}".format(Submodel1.__module__, Submodel1.__name__).encode("utf-8"),
     ],
 )
 def test_simple_model(sheraf_database, model):
     sheraf_database.reset()
 
     class Other(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     with sheraf.connection() as conn:
-        a = Submodel.create()
+        a = Submodel1.create()
 
         other = Other.create()
         assert other.submodel is None
-
         other.submodel = a
+        assert a.id == other.mapping["submodel"]
 
         _other = Other.read(other.id)
         assert a == _other.submodel
@@ -47,10 +53,10 @@ def test_simple_model(sheraf_database, model):
 
 def test_set_to_none(sheraf_connection):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     model = Model.create()
-    model.submodel = Submodel.create()
+    model.submodel = Submodel1.create()
     assert model.submodel
     model.submodel = None
     assert model.submodel is None
@@ -58,17 +64,17 @@ def test_set_to_none(sheraf_connection):
 
 def test_create(sheraf_connection):
     class ModelForTest(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     model = ModelForTest.create(submodel={"name": "A"})
-    assert isinstance(model.submodel, Submodel)
+    assert isinstance(model.submodel, Submodel1)
     assert isinstance(model.submodel.mapping, sheraf.types.SmallDict)
     assert "A" == model.submodel.name
 
 
 def test_update_edition(sheraf_database):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     with sheraf.connection(commit=True):
         model = Model.create(submodel={"name": "YOH"})
@@ -77,21 +83,21 @@ def test_update_edition(sheraf_database):
     with sheraf.connection(commit=True):
         model.edit(value={"submodel": {"name": "YEAH"}}, edition=True)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YEAH" == Model.read(model.id).submodel.name
         assert model.submodel.id == last_sub_id
 
     with sheraf.connection():
         model = Model.read(model.id)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YEAH" == Model.read(model.id).submodel.name
         assert model.submodel.id == last_sub_id
 
 
 def test_update_no_edition(sheraf_database):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     with sheraf.connection(commit=True):
         model = Model.create(submodel={"name": "YOH"})
@@ -100,21 +106,21 @@ def test_update_no_edition(sheraf_database):
     with sheraf.connection(commit=True):
         model.edit(value={"submodel": {"name": "YEAH"}}, edition=False)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YOH" == Model.read(model.id).submodel.name
         assert model.submodel.id == last_sub_id
 
     with sheraf.connection():
         model = Model.read(model.id)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YOH" == Model.read(model.id).submodel.name
         assert model.submodel.id == last_sub_id
 
 
 def test_update_replacement(sheraf_database):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     with sheraf.connection(commit=True):
         model = Model.create(submodel={"name": "YOH"})
@@ -123,21 +129,21 @@ def test_update_replacement(sheraf_database):
     with sheraf.connection(commit=True):
         model.edit(value={"submodel": {"name": "YEAH"}}, replacement=True)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YEAH" == Model.read(model.id).submodel.name
         assert model.submodel.id != last_sub_id
 
     with sheraf.connection():
         model = Model.read(model.id)
 
-        assert isinstance(Model.read(model.id).submodel, Submodel)
+        assert isinstance(Model.read(model.id).submodel, Submodel1)
         assert "YEAH" == Model.read(model.id).submodel.name
         assert model.submodel.id != last_sub_id
 
 
 def test_delete_before_created(sheraf_database):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel)
+        submodel = sheraf.ModelAttribute(Submodel1)
 
     with sheraf.connection(commit=True):
         m = Model.create()
@@ -146,8 +152,50 @@ def test_delete_before_created(sheraf_database):
 
 def test_indexation(sheraf_connection):
     class Model(tests.UUIDAutoModel):
-        submodel = sheraf.ModelAttribute(Submodel).index()
+        submodel = sheraf.ModelAttribute(Submodel1).index()
 
-    s = Submodel.create(name="foo")
+    s = Submodel1.create(name="foo")
     m = Model.create(submodel=s)
     assert [m] == Model.search(submodel=s)
+
+
+def test_generic_nominal_case(sheraf_database):
+    class Model(tests.UUIDAutoModel):
+        submodel = sheraf.ModelAttribute((Submodel1, Submodel2))
+
+    with sheraf.connection(commit=True):
+        s1 = Submodel1.create(name="foo")
+        s2 = Submodel2.create(name="bar")
+
+        m1 = Model.create(submodel=s1)
+        m2 = Model.create(submodel=s2)
+
+        assert (Submodel1.table, s1.id) == m1.mapping["submodel"]
+        assert (Submodel2.table, s2.id) == m2.mapping["submodel"]
+
+        assert "foo" == m1.submodel.name
+        assert "bar" == m2.submodel.name
+
+    with sheraf.connection():
+        m1 = Model.read(m1.id)
+        m2 = Model.read(m2.id)
+
+        assert isinstance(m1.submodel, Submodel1)
+        assert isinstance(m2.submodel, Submodel2)
+
+        assert "foo" == m1.submodel.name
+        assert "bar" == m2.submodel.name
+
+
+def test_generic_indexation(sheraf_connection):
+    class Model(tests.UUIDAutoModel):
+        submodel = sheraf.ModelAttribute((Submodel1, Submodel2)).index()
+
+    s1 = Submodel1.create(name="foo")
+    s2 = Submodel2.create(name="bar")
+
+    m1 = Model.create(submodel=s1)
+    m2 = Model.create(submodel=s2)
+
+    assert [m1] == Model.search(submodel=s1)
+    assert [m2] == Model.search(submodel=s2)